requirements.txt
pandas>=2.0.0
yfinance>=0.2.40
matplotlib>=3.7.0
python-dotenv>=1.0.0
